added transaction and formatter classes
these form the core of how we interact with the data. Transactions represent the relevant parts of each row in the .csv, each as a field in a dictionary called "properties". Formatters contain a list of such transactions, and so can perform "meta operations" on these transactions, such as sorting them or using the range of values that, for example, Anon_StudentId takes.